Refactor the APIService into smaller parts
Having a single monolithic APIService is going to get clunky so split out dedicated services for each bit - notes, clubs etc. -  that inherit from a base service that can deal with the common bits of executing an API request to get a JSON response or build fully formed URLs from a local fragment.