Adjust how scrolling is handled on notes page so that only the notes list is scrollable and can remove the position sticky bodge from the filer
The notes list is contained within an absolutely positioned div that's stretched to fit the size of the parent positioned element. Then we can enable scrolling on that list div and have only that part of the page scroll and keep everything else in place. Still need to restore the loading overlay