Batch notes page state that is changed together into a single state object that can be updated in one go to avoid additional renders
Loading the data for the notes page is triggering several renders due to having multiple separate state variables for things like the notes, tags, loading state etc. After fetching the data from the server each one is updated in turn which triggers a render. By combining those variables into a single state object we can cut down on the amount of work needed to initially render the page.

When testing this a bug was found where deleting notes out of order was slightly wonkly - for example given notes 1, 2, 3, deleting note 2 worked OK, but deleting note 3 would show the info for note 2 in the prompt and not actually refresh the UI after it was deleted. This is because we create a react-spring trail for everything in the notes list but when we delete stuff, we soft-delete it so that it's animated out of view so the trail and subsequent animated.div list contained stuff for soft-deleted notes which threw off the note item bound to each note list item.

To fix this we filter out soft-deleted notes and only use those in useTrail so we only render visible notes. When binding each note list item we then use the index from .map() to lookup in the visible notes collection rather than the notes collection from within the current context