Add a small delay when deleting things before forwarding the delete event.
This gives the super spiffy delete button animation time to show rather than it showing and instantly disappearing. The approach here is a bit hacky as it simply uses setTimeout rather than hooking into animationend events but trying to use animationend is a PITA since the button is a dxButton so we'd have to go digging about for the DOM element we want to listen to.