Replace Validator usage with a home grown validation solution
There seems to be a bub with the Validate/RequiredRule etc. in that despite resetting it on close, the next time the modal is shown the indicators for missing content is still shown which is not what we want. Instead, revert back to how we used to do it by manually checking that the required stuff is entered