Avoid a whole bunch of useless re-renders on the notes page
Re-organise the components on the page to avoid a bunch of re-renders - for example when toggling the sidebar, we'd re-render the notes list which is unnecessary and wasteful. By adding in a new NotesPageContainer component whcih contains the sidebar and the toolbar options, we can embed the NotesList component as a child of NotesPageContainer to take advantage of that when React is handed the same chunk of JSX for a component as the previous render, it doesn't need to traverse that component tree to determine if a re-render is necessary.

Inspiration for this comes from Dan Abramov's blog post https://overreacted.io/before-you-memo/