Mostly fix a layout issue with the notes list when adding/removing tags
When editing an existing note without tags to include tags, the height of the list item wasn't being increased to reflect the new content so would overlap the next one along. Likewise removing all tags from a note list item would leave a big gap as the height wasn't reduced accordingly.

Fixed by using scrollHeight when measuring the list item's height which will give the height of the content, not the element containing it. In this scenario this allows for the height of the list item to be animated out to fit.

The changes to the JSX are needed so that we get the correct scrollHeight calculated, otherwise when removing all tags, the height wasn't decreasing as we measure the height on the outermost div which gave the wrong answer. Now the outermost div has the actual content within it so we can get the right height