Use react-query to handle managing talking to the backend API service
This has turned out to have several benefits:
 - Remove a bulk of the stuff currently in the notes context into custom hooks which wrap react-query.
- These new custom hooks also contain logic that was previously in notes-page-controller which was always a little odd. Having this logic in hooks feels more "React"-y way of doing things.
- We can use react-query's built in bits for tracking whether data is being fetched or not to show/hide the loading overlay. This component needed some tweaks to make it fit the react-query API so rather than it wrapping a bunch of child components, it's simply a component that's shown on it's own and will be hidden when the data fetch is completed.

The only stuff left in the page controller/context is related to filtering. This is currently commented out until that's transition to another custom hook to handle filtering.

This change also includes some bug fixes for how callbacks were handled on the delete buttons as we wasn't passing back the note to be deleted